Added the ability to specify the auth header and refactored LocalFile
Currently the auth header is specified for requests to Stripe (this is derived from the secret key provided and is necessary). However the same auth header is also used when fetching images, which can be hosted remotely. This can actually cause the download to fail and return a 401 Unauthorized. You should now be able to provide custom auth credentials which will be added to the Authorization header field of the HTTP request. If no credentials are provided then there will be no Authorization header on the HTTP request.

As part of this commit I have also heavily refactored the LocalFile file to increase readability, as well as fixed the problem with files not having finished downloading by the time sourceNodes() returned.