Remove auth option, fixed file download, and refactored LocalFile
We had previously made the distinction between files hosted on Stripe and those that weren't because we assumed that all files hosted on Stripe needed authentication. We now know that some Files hosted on Stripe are public while others are private, which makes this distinction unnecessary. Now we have a single method that allows you to specify whether you would like to authenticate or not. 

Also fixed an issue with the convertToArray method where it incorrectly converted an object to an empty array.

Finally removed the auth option because it is currently redundant (we know which files do and don't need authentication thanks to Stripes docs).